==> scripts/03_ev_adoption.ipynb <==
import pandas as pd 
import numpy as np
import os


Sources: 
EV data:
https://www.iea.org/data-and-statistics/data-tools/global-ev-data-explorer

Total cars:
https://worldpopulationreview.com/country-rankings/cars-by-country



# Set up workspace path - go up one directory from scripts folder to project root
workspace_root = os.path.dirname(os.getcwd())

# Set the data directory path
data_dir = os.path.join(workspace_root, 'data')

EV_cars_path = os.path.join(data_dir, 'cars_country.csv')
cars_country_path = os.path.join(data_dir, 'cars-by-country-2025.csv')
vars_WDI = os.path.join(data_dir, 'important_variables.csv')


ev_cars = pd.read_csv(EV_cars_path)
cars_country = pd.read_csv(cars_country_path)
vars_WDI = pd.read_csv(vars_WDI)[[
    'REF_AREA_LABEL', 
    'YEAR', 
    'WB_WDI_EN_GHG_CO2_PC_CE_AR5', #Carbon dioxide (CO2) emissions excluding LULUCF per capita (t CO2e/capita)
    'WB_WDI_EG_FEC_RNEW_ZS', # Renewable energy consumption (% of total final energy consumption)
    'WB_WDI_EG_USE_PCAP_KG_OE', # Energy use (kg of oil equivalent per capita)
    'WB_WDI_NY_GDP_MKTP_KD_ZG', #	GDP growth (annual %)
    'WB_WDI_SP_URB_GROW', # Urban population growth (annual %)
    'WB_WDI_EG_USE_ELEC_KH_PC'	#Electric power consumption (kWh per capita)
]]


I chose 2022 as is the most recent year with range available data

ev_cars_2024 =ev_cars.groupby(['country','year']).agg({'value': 'sum'}).sort_values('year', ascending=False).reset_index()

ev_cars_2024 = ev_cars_2024[ev_cars_2024['year'] == 2022]
ev_cars_2024.rename(columns={'value': 'stock_ev'}, inplace=True)

ev_cars_2024['country'] = ev_cars_2024['country'].replace('USA', 'United States')
# Stock of EV
ev_cars_2024


cars_2022 = cars_country[cars_country['CarsByCountryDataYear'] == 2022][['country', 'CarsByCountryTotal', 'CarsByCountryDataYear']]
cars_2022.rename(columns={'CarsByCountryTotal': 'total_cars'}, inplace=True)
cars_2022.rename(columns={'CarsByCountryDataYear': 'year'}, inplace=True)
cars_2022


vars_WDI = vars_WDI[vars_WDI['YEAR'] == 2022]
vars_WDI.rename(columns={'REF_AREA_LABEL': 'country'}, inplace=True)
vars_WDI.rename(columns={'YEAR': 'year'}, inplace=True)
vars_WDI

merged = pd.merge(vars_WDI, cars_2022, on=['country', 'year'], how='inner')

# Fusionar con ev_cars_2024
final_merged = pd.merge(merged, ev_cars_2024, on=['country', 'year'], how='inner')

final_merged['ev_adoption_rate'] = (final_merged['stock_ev'] / final_merged['total_cars']) * 100
final_merged

### Defininig a model

# Data

features = [
    'ev_adoption_rate',
    'total_cars',
    'WB_WDI_EG_FEC_RNEW_ZS',
    'WB_WDI_EG_USE_PCAP_KG_OE',
    'WB_WDI_EG_USE_ELEC_KH_PC',
    'WB_WDI_NY_GDP_MKTP_KD_ZG',
    'WB_WDI_SP_URB_GROW'
]

X = final_merged[features]
y = final_merged['WB_WDI_EN_GHG_CO2_PC_CE_AR5']



# Model 
from sklearn.metrics import r2_score, mean_squared_error
from sklearn.linear_model import LinearRegression
import numpy as np

# Entrenar
model = LinearRegression()
model.fit(X, y)

# Predicciones
y_pred = model.predict(X)

# Métricas
r2 = r2_score(y, y_pred)
rmse = np.sqrt(mean_squared_error(y, y_pred))

print(f"R²: {r2:.4f}")
print(f"RMSE: {rmse:.4f}")


# Interpretation
coef_df = pd.DataFrame({
    'Variable': features,
    'Coeficiente': model.coef_
}).sort_values(by='Coeficiente', key=abs, ascending=False)

print(coef_df)

### Simulation of adoption

simulated = final_merged.copy()
simulated['ev_adoption_rate'] = 0.50  # 50% adopción EV

# Predicciones
y_simulated = model.predict(simulated[features])

# Comparación
final_merged['pred_base'] = y_pred
final_merged['pred_scenario'] = y_simulated
final_merged['change_abs'] = final_merged['pred_scenario'] - final_merged['pred_base']
final_merged['change_pct'] = 100 * final_merged['change_abs'] / final_merged['pred_base']

final_merged[['country','year','pred_base','pred_scenario','change_abs','change_pct']]
